requirements.txt
python-dotenv
sqlmodel
psycopg[binary]
typer